01_CYBER_BULLYING_ON_TWITTER/app/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.1
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
filelock==3.12.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
packaging==23.1
pathy==0.10.1
preshed==3.0.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
requests==2.30.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.12
srsly==2.4.6
sympy==1.11.1
thinc==8.1.10
torch==2.0.0
torchdata==0.6.0
torchtext==0.15.1
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==2.0.2
wasabi==1.1.1
Werkzeug==2.3.3
zipp==3.15.0
